Refactoring the whole project (marc_to_wikidata) into multiple files.
