+++ Util.py
@@ -1,0 +1,7 @@
+language_map = {
+    'ara': 'ar',
+    'cyr': 'ru',
+    'fre': 'fr',
+    'heb': 'he',
+    'lat': 'en',
+}
