+++ storeInWikidata.py
@@ -1,0 +1,3 @@
+def create_new_record_in_wikidata(record):
+    print("Not Implemented!")
+    # raise NotImplemented
